Force clearwater-infrastructure to run after cloud-init.
On servers using systemd as their init system, systemd generates unit
information for legacy sysvinit-style scripts based on the LSB headers
in the scripts. It uses the dependencies in the scripts to ensure that
legacy services are started at the right point.

If clearwater-infrastructure is used without clearwater-startup, on a
system using cloud-init then it's important that clearwater-infrastructure is
started after cloud-init has finished, so that cloud-init doesn't
overwrite the configuration set up by clearwater-infrastructure, in
particular the hostname of the system as defined by the field
`public_hostname` in /etc/clearwater/local_config.

Adding this header should not affect systems running upstart as an init
system (e.g. Ubuntu 14.04), because upstart uses the runlevel to
trigger legacy init scripts rather than building a dependency graph
like systemd.